backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
cached-property==1.5.2
click==7.1.2
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-url-filter==0.3.15
djangorestframework==3.12.1
enum-compat==0.0.3
mypy-extensions==0.4.3
pathspec==0.8.0
pytz==2020.1
regex==2020.10.28
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
yapf==0.30.0
